# projekat1/gateway/static/openapi.yml
openapi: 3.0.0

info:
  version: 1.0.0
  title: Gateway API 
  description: API  se koristi u prvom i drugom projektu

servers:
  - url: http://localhost:5000
    description: Lokalni server

paths:
  /:
    get:
      summary: Vrati podatke o filmu ciji id se prosledjuje
      description: "Jedini endpoint bitan za Projekat 2.<br>Samo naslove trazenih filmova Gateway prosledjuje Analytics mikroservisu.<br>Ideja je da se registruju najtrazeniji filmovi u kratkom vremenskom periodu.<br>Vise o tome u README.md: '5. ekuiper' i '8. analytics'. "
      operationId: get
      tags:
        - movies
      parameters:
        - name: _id
          in: query
          schema: 
            type: string
            minLength: 24
            maxLength: 24
          required: true
          example: '63b449736b010eff5124fa04'
          description: movie id unutrasnjeg API-a.<br><b>Example ID nije validan!</b><br>Za validan ID izabari neki od filmova dobijenih <i>/suggestions</i> zahtevom ili ID dobijen <i>/create</i> zahtevom.
      responses:
        '200':
          description: Detalji o trazenom filmu
          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/Movie' 
        '204':
          description: No Content
        '400': 
          description: Bad Request
  /create:
    post:
      summary: Dodaj podatke o novom filmu
      operationId: create
      tags: 
        - movies
      requestBody:
        description: "Pogledaj semu za obavezna polja. Polja koja nisu obavezna i cija je vrednost nepoznata ne treba navoditi."
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MovieCreation'
      responses:
        '201':
          description: "Film kreiran.<br>Vraca ID kreiranog filma.<br>Podaci se mogu proveriti metodom GET."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdObject'
        '400':
          description: Bad Request
        '500':
          description: Internal Server Error
  /update:
    put:
      summary: Prepravi podatke o postojecem filmu.
      operationId: update
      tags:
        - movies
      requestBody: 
        description: "Nacin rada je nepraktican i podlozan greskama, ali u korist laksoj implementaciji. <br> Princip rada: <br> 1. ID filma koji se referencira mora da postoji u bazi. <br> 2. Sva polja iz primera moraju biti popunjena. <br> 3. Polja koja se ne menjaju treba da imaju stare vrednosti u formatu kao za kreiranje novog filma. <br> 3.1 PRIMER: GET ce za 'cast' vratiti jedan string u kome su glumci odvojeni zarezom. Ukoliko ne zelimo da menjamo 'cast', ali koristimo ovaj endpoint, onda je neophodno da se taj jedan string izdeli i vrati kao niz stringova.<br> 4. Polja koja se menjaju sadrze samo konacnu vrednost, vrednost kakva ce ostati u bazi.<br>5. Proveri unete podatke GET metodom."
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MovieUpdate'
      responses:
        '200':
          description: 'OK'
        '400':
          description: Bad Request
        '500':
          description: Internal Server Error
  /delete:
    delete:
      summary: Obrisi podatke o filmu sa zadatim ID-em
      operationId: delete
      tags:
       - movies
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IdObject'
      responses:
        '200':
          description: 'OK'
        '400':
          description: 'Bad Request'
  /suggestions:
    post:
      summary: 'Daje sugestije na osnovu poslatih parametara'
      operationId: suggestions
      tags:
        - movies
      requestBody:
        description: 'Body mora da sadrzi barem jedan neprazan parametar.<br>Slanje parametra koji se ne poklapa sa semom nije validan.'
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FilterObject'
      responses:
        '200':
          description: 'OK<br>Ocekuje se lista filmova koji ispunjavaju uslove "pretrage".'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Movie'
        '400':
          description: 'Bad Request'
            


components:
  schemas:
    Movie:
      type: object
      properties:
        _id:
          type: string
        cast: 
          type: string
          description: CSV cuva podatke u jednom stringu umesto u nizu
        country:
          type: string
        date_added: 
          type: string
        description:
          type: string
        director:
          type: string
        duration:
          type: string
        listed_in:
          type: string
          description: Zanrovi
        rating:
          type: string
        release_year:
          type: integer
        score_average:
          type: integer
          description: Srednji rejting
        title:
          type: string
        trakt:
          type: string
          description: Link do informacija o flmu na sajtu trakt.tv
        imdb:
          type: string
          description: Link do informacija o flmu na sajtu imdb.com
        type:
            type: string
            description: Uvek vraca Movie
      example:
        _id: "638c90fac44a1fb620ab2402"
        cast: "Melissa McCarthy, Chris O'Dowd, Kevin Kline, Timothy Olyphant, Daveed Diggs, Skyler Gisondo, Laura Harrier, Rosalind Chao, Kimberly Quinn, Loretta Devine, Ravi Kapoor"
        country: "United States"
        date_added: "September 24, 2021"
        description: "A woman adjusting to life after a loss contends with a feisty bird that's taken over her garden — and a husband who's struggling to find a way forward."
        director: "Theodore Melfi"
        duration: "104 min"
        listed_in: "Comedies, Dramas"
        rating: "PG-13"
        release_year: 2021
        score_average: 56
        title: "The Starling"
        imdb: www.imdb.com/title/tt5164438/
        trakt: trakt.tv/movies/the-starling-2021
        type: "Movie"

    StringArray:
      type: array
      items: 
        type: string
        
    NotEmptyStringArray:
      type: array
      items:
        type: string
        minLength: 1
        
    IdObject:
      type: object
      properties:
        _id: 
          type: string
          example: '<valid movie id>'

    MovieCreation:
      type: object
      required:
        - duration
        - title
        - release_year
        - director
      properties:
        duration:
          type: integer
          format: int32
        title:
          type: string
        release_year:
          type: integer
          format: int64
        director:
          $ref: '#/components/schemas/StringArray' 
        cast:
          $ref: '#/components/schemas/StringArray' 
        country:
          $ref: '#/components/schemas/StringArray' 
        rating:
          type: string
        listed_in:
          $ref: '#/components/schemas/StringArray' 
        description:
          type: string
      example:
        duration: 72
        title: 'Prokleta avlija'
        release_year: 1984
        director: 
          - 'Ivo Andric'
          - 'Milenko Maricic'
        cast:
          - 'Petar Kralj'
          - 'Dubravko Jovanovic'
        country: 
          - 'Serbia'
          - 'Bosnia and Herzegovina'
        rating: 'PG-13'
        listed_in:
          - 'Drama'
          - 'Psychological'
        description: 'Bosnian Franciscan monk is put into an infamous prison in Istanbul. There he witnesses the sad destiny of a young Ottoman scholar.'
    
    MovieUpdate:
      type: object
      required:
        - _id
        - duration
        - title
        - release_year
        - director
        - cast
        - country
        - rating
        - listed_in
        - description
      properties:
        _id:
          type: string
          minLength: 24
          maxLength: 24
        duration:
          type: integer
          format: int32
        title:
          type: string
        release_year:
          type: integer
          format: int64
        director:
          $ref: '#/components/schemas/StringArray' 
        cast:
          $ref: '#/components/schemas/StringArray' 
        country:
          $ref: '#/components/schemas/StringArray' 
        rating:
          type: string
        listed_in:
          $ref: '#/components/schemas/StringArray' 
        description:
          type: string
      example:
        _id: '63b18d4b9e5b5f62217a85cd'
        duration: 72
        title: 'Prokleta avlija'
        release_year: 1984
        director: 
          - 'Ivo Andric'
          - 'Milenko Maricic'
        cast:
          - 'Petar Kralj'
          - 'Dubravko Jovanovic'
        country: 
          - 'Serbia'
          - 'Bosnia and Herzegovina'
        rating: 'PG-13'
        listed_in:
          - 'Dramas'
        description: 'Novi opis bez progutanih slova.'
        
    FilterObject:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
          minLength: 1
        director: 
          $ref: '#/components/schemas/NotEmptyStringArray'
        cast:
          $ref: '#/components/schemas/NotEmptyStringArray'
        listed_in:
          $ref: '#/components/schemas/NotEmptyStringArray'
      example:
        title: 'man'
        listed_in: 
          - 'Dramas'
          - 'Comedies'